feat(web): 사이드바 내비 통폐합 — 24→15 항목 (허브 탭 + 설정 흡수)
사이드 내비 24항목(관리자 기준)을 15항목으로 재편. 묶인 페이지들은 기존
라우트를 유지한 채 PageHeader 아래 탭 바(PageTabs)로 오간다.

Phase 1 — nav.ts 재편:
- 통합 그룹: MCP 단일 항목(/mcp-servers)
- MCP 모니터링(데이터 소스가 전부 /api/admin/* — 일반 유저에겐 빈 화면이던
  배치 결함)·에이전트 학습을 관리자 관측 허브로 이동
- 계정 그룹: 설정 / 사용량 / 개발자 3항목

Phase 2 — 허브 탭 (components/hub-tabs.tsx + ui/page-tabs.tsx):
- MCP: 내 서버 | 카탈로그
- 개발자: 문서 | API 액세스

Phase 3 — 설정 흡수:
- 메모리 → 설정 '메모리' 탭 (components/settings/memory-section.tsx)
- API 키(외부 LLM BYOK) → 설정 '모델' 탭 하단
  (components/settings/provider-keys-section.tsx). 모델 탭 내 /api-keys
  링크는 같은 탭에 섹션이 내려앉아 텍스트 안내로 대체
- 설정이 ?tab= 쿼리로 초기 탭 결정 (useSearchParams — 동적 라우트)
- 구 라우트 딥링크 호환: /memory → /settings?tab=memory,
  /api-keys → /settings?tab=model (redirect)

Phase 4 — 관리자 관측 허브:
- 대시보드 | 애널리틱스 | 메트릭 | MCP 모니터링 | 에이전트 학습 5탭

i18n: nav.items.mcp/developer, pageTabs.*, settings.tabs.memory 4언어 추가
(기존 번역 재활용). 모바일 하단 탭바는 유지 라우트만 참조라 무변경.

검증: tsc·eslint 0건, next build 성공, 임시 서버(:3105)에서 Playwright
라이브 확인 — 관리자(실 admin 토큰): 사이드바 15항목·허브 탭 3종·설정
메모리 탭·모델 탭 실DB 키 2건 렌더·redirect 2종 / 게스트: 채팅만 노출.
