+++ apps/web/app/(workspace)/admin/metrics/page.tsx
@@ -18,6 +18,7 @@ import {
   Td,
 } from "@/components/ui/primitives";
 import { ApiClient } from "@/lib/api-client";
+import { AdminObservabilityTabs } from "@/components/hub-tabs";
 
 type NodeStatus = "online" | "degraded" | "offline";
 
@@ -191,6 +192,7 @@ export default function AdminMetricsPage() {
           </Button>
         }
       />
+      <AdminObservabilityTabs />
 
       <div className="min-h-0 flex-1 overflow-y-auto p-6">
         <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 lg:grid-cols-4">
